requirements.txt
Django==4.0  # django-celery-beat: Django <= 4.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-cleanup==6.0.0
django-templated-mail==1.1.1
django-cloudinary-storage==0.3.0
django-celery-email==3.0.0
django-celery-beat==2.3.0
django-pickling==1.0
django-redis==5.2.0
django-cacheops==6.1

djangorestframework==3.13.1
drf-spectacular==0.23.1

orjson==3.7.8
psycopg2-binary==2.9.3
redis==4.3.4
Pillow==9.2.0
gunicorn==20.1.0
whitenoise==6.2.0
celery==5.2.7
gevent==21.12.0
argon2-cffi==21.3.0
validated-dc==1.3.3
pydantic[email]==1.9.2

# dev
django-silk==5.0.1

# test
parameterized==0.8.1
factory_boy==3.2.1
Faker==14.1
